front/requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.38.0
streamlit-feedback==0.1.3
pillow==10.4.0
PyYAML==6.0.1
requests==2.31.0
typing_extensions==4.12.2
pydantic==2.10.6
